+++ ingestion/sonar-project.properties
@@ -3,7 +3,7 @@ sonar.projectName=open-metadata-ingestion
 sonar.organization=open-metadata
 sonar.language=py
 
-sonar.sources=./src/metadata
+sonar.sources=src/metadata
 sonar.tests=tests
 sonar.exclusions=src/metadata_server/static/**,ingestion/src/metadata_server/templates/**
 sonar.python.xunit.reportPath=junit/test-results-*.xml
