+++ ingestion/src/metadata/utils/singleton.py
@@ -1,0 +1,14 @@
+from abc import ABCMeta
+
+
+class Singleton(ABCMeta):
+    _instances = {}
+
+    def __call__(cls, *args, **kwargs):
+        if cls not in cls._instances:
+            cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
+        return cls._instances[cls]
+
+    @classmethod
+    def clear_all(cls):
+        Singleton._instances = {}
